First communication successfully build! - The client is now synchronized with the core library (finally), - Packet reading errors have been fixed, - Added new configurations for managing starter packets, - Client outgoing channels have been improved.

+++ RMF.Core/Events/Client/HeartbeatEvent.cs
@@ -21,7 +21,7 @@ namespace RMF.Core.Events.Client
             {
                 heartbeatPacket.Timestamp = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds();
                 session.SendPacket(heartbeatPacket);
-                await Task.Delay((int)(this.IntervalSecs * 1000), token);
+                await Task.Delay(this.IntervalSecs * 1000, token);
             }
         }
     }
